Add files via upload
This is a SQL query for my portfolio. The premise is that I work for a company that owns a cruise line and airline. They want to do a marketing campaign to get returning customers. I have a dataset of customers and I want to advertise the closest cruise port and airport that meets the customer's cruise preference. I also only want to advertise air travel if the customer is more than 250 miles from the airport. Otherwise, the customer is sent to our partnered rental car company. With this SQL query I am able to isolate the customer preference, then measure the distance to the closest port and airport so marketing is equipped with the information they need to execute the marketing campaign.