BigQuery SQL Example
This is a query I wrote for BigQuery. The goal was to display the rank changes on an WIP record that is being worked by stakeholders. This query sets up the data to track the rank changes and ranked by person and instance of interaction with a person at any given time. We created a power BI report to evaluate the ranking program as a whole.