BigQuery Ranking & Lag Examples
This is a query that was used to find various rankings in a set of data containing interactions with people. I ranked the interactions with each person based on the person as a whole and within each interaction as the interaction record data changed. There is a record recorded every time something in the interaction changed. I used the lag function to highlight when the interactions changed the actual ranking rather than just minor changes that don't effect rank.